tests/requirements.txt
grpcio>=1.68.0
grpcio-tools>=1.68.0
protobuf>=5.29.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
